fix: use latest js-ipfs and ipfs-http-client
It was impossible to build old versions anymore. This switches to the
latest js-ipfs. It also switched to ipld-explorer-components@2.0.4 which
works around breaking changes in the Core JS API.

In short, we now can use the `ipfs.block.get` from the latest js-ipfs
without the need for rewriting entire IPLD Explorer.